+++ common/src/main/java/io/github/flemmli97/improvedmobs/events/EventCalls.java
@@ -21,7 +21,6 @@ import io.github.flemmli97.tenshilib.platform.registry.RegistryHelper;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
 import net.minecraft.server.level.ServerLevel;
-import net.minecraft.world.Container;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.entity.Entity;
@@ -159,7 +158,7 @@ public class EventCalls {
         }
         boolean villager = !Config.CommonConfig.entityBlacklist.hasFlag(mob, EntityModifyFlagConfig.Flags.TARGETVILLAGER, Config.CommonConfig.targetVillagerWhitelist);
         boolean aggressive;
-        if ((mob instanceof NeutralMob)) {
+        if ((mob instanceof NeutralMob) && !Config.CommonConfig.entityBlacklist.hasFlag(mob, EntityModifyFlagConfig.Flags.NEUTRALAGGRO, Config.CommonConfig.neutralAggroWhitelist)) {
             aggressive = Config.CommonConfig.neutralAggressiv != 0 && mob.level.random.nextFloat() <= Config.CommonConfig.neutralAggressiv;
             if (aggressive)
                 mob.targetSelector.addGoal(1, setNoLoS(mob, Player.class, flags.canBreakBlocks == EntityFlags.FlagType.TRUE || mob.level.random.nextFloat() < 0.5, null));